// types/job.ts
export type Job = {
  id: string;
  title: string;
  description: string;
  company: string;
  location: string;
  salary_from: number;
  salary_to: number;
  employment_type: string;
  application_deadline: string;
  qualifications: string[] | string;
  contact: string;
  job_category: string;
  is_remote_work: number;
  number_of_opening: number;
  created_at: string;
  updated_at: string;
};